1- fix bugs

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include "function_pointer.h"
+#include "function_pointers.h"
 
 /**
  * array_iterator - executes a function given as a parameter
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	long unsigned int i;
 
 	if (array != NULL && size > 0 && action != NULL)
 	{
